ConfigParser: copy constructor and getters set up

+++ inc/ConfigParser.hpp
@@ -21,7 +21,7 @@ public:
 
 	void	output(void);
 
-	int findClosingBracket(int i, size_t pos);
+	int findClosingBracket(size_t i, size_t pos);
 
 	string&	operator[](int i);
